requirements.txt
line-bot-sdk
flask
snownlp
hanziconv
